% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity.R
\name{activity}
\alias{activity}
\title{Activity}
\usage{
activity(activity, day_of_week, friend_name)
}
\arguments{
\item{activity}{a character for activity. Verbs work best}

\item{day_of_week}{a character for day of the week you will schedule the activity}

\item{friend_name}{a character for the name of the friend you will do the activity with}
}
\value{

}
\description{
Activity
}
