% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pizza.R
\name{pizza}
\alias{pizza}
\title{Pizza Order}
\usage{
pizza(quantity, size, type)
}
\arguments{
\item{quantity}{a number representing the number of pizzas you would like to order}

\item{size}{a character for the size of pizza you want}

\item{type}{a character for the type of pizza you want}
}
\value{

}
\description{
Pizza Order
}
