% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sundae.R
\name{sundae}
\alias{sundae}
\title{Sundae Order}
\usage{
sundae(flavor, scoops, container)
}
\arguments{
\item{flavor}{a character containing the input ice cream flavor}

\item{scoops}{a number refering the the number of scoops in the sundae}

\item{container}{a character that reflects the container that the sundae should be served from}
}
\value{

}
\description{
contains a function that creates a sundae order
}
